refactor use command
